Make cutout in the decorated room avatar transparent rather than fixed
