--- res/css/views/rooms/_RoomTileIcon.scss
@@ -1,69 +1,0 @@
-/*
-Copyright 2020 The Matrix.org Foundation C.I.C.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-.mx_RoomTileIcon {
-    width: 12px;
-    height: 12px;
-    border-radius: 12px;
-    background-color: $roomlist-bg-color; // to match the room list itself
-}
-
-.mx_RoomTileIcon_globe::before {
-    content: '';
-    width: 8px;
-    height: 8px;
-    top: 2px;
-    left: 2px;
-    position: absolute;
-    mask-position: center;
-    mask-size: contain;
-    mask-repeat: no-repeat;
-    background: $secondary-fg-color;
-    mask-image: url('$(res)/img/globe.svg');
-}
-
-.mx_RoomTileIcon_offline::before {
-    content: '';
-    width: 8px;
-    height: 8px;
-    top: 2px;
-    left: 2px;
-    position: absolute;
-    border-radius: 8px;
-    background-color: $presence-offline;
-}
-
-.mx_RoomTileIcon_online::before {
-    content: '';
-    width: 8px;
-    height: 8px;
-    top: 2px;
-    left: 2px;
-    position: absolute;
-    border-radius: 8px;
-    background-color: $presence-online;
-}
-
-.mx_RoomTileIcon_away::before {
-    content: '';
-    width: 8px;
-    height: 8px;
-    top: 2px;
-    left: 2px;
-    position: absolute;
-    border-radius: 8px;
-    background-color: $presence-away;
-}
